fix - current sensor

+++ main/modules/currentSensor/currentSensor.h
@@ -5,12 +5,12 @@ class CurrentSensor
 {
 public:
     CurrentSensor(int CURRpin);
-    int getCurrent();
+    float getCurrent(); // Change to float to allow for fractional current values
 
 private:
     int _CURRpin;
     const float sensorSensitivity = 100.0; // Sensitivity of ACS712 (e.g., 100mV per A for ACS712-05B)
-    const float zeroCurrentVoltage = 2.5;
+    const float zeroCurrentVoltage = 2.5;  // Zero current voltage in V
 };
 
-#endif
+#endif // CURRENTSENSOR_H
